singly_linked_list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while(h != NULL)
 	{
@@ -20,5 +20,4 @@ size_t print_list(const list_t *h)
 		h = h->next;
 	}
 	return (count);
-}
-		
+}
